Updated the login screen and added the forgot password screen
Updated the design layout of the login screen to match the sign up screen so the it looks more cohesive. Implemented firebase authentication to the login page, ensuring that the user is inputting a valid email and password associated with an account. Added the forgot password screen and used firebase to send the email link to reset the password associated with the specified email account