Updated notification layout and added documentation
Updated the dummy data for the notification layouts to fill the screen and added another section for notifications that are older than 30 days.
Changed from using a FlatList to a SectionList to add section headers for the different time categories.
Added documentation/comments in the Login Screen and Forgot Password Screen, and the Notification Screen